Refactors elastic extensions to higher level
