+++ src/main/kotlin/com/amazon/opendistroforelasticsearch/indexmanagement/indexstatemanagement/elasticapi/ElasticExtensions.kt
@@ -20,99 +20,7 @@ package com.amazon.opendistroforelasticsearch.indexmanagement.indexstatemanageme
 import com.amazon.opendistroforelasticsearch.indexmanagement.indexstatemanagement.model.ManagedIndexMetaData
 import com.amazon.opendistroforelasticsearch.indexmanagement.indexstatemanagement.model.coordinator.ClusterStateManagedIndexConfig
 import com.amazon.opendistroforelasticsearch.indexmanagement.indexstatemanagement.settings.ManagedIndexSettings
-import com.amazon.opendistroforelasticsearch.jobscheduler.spi.utils.LockService
-import kotlinx.coroutines.delay
-import org.apache.logging.log4j.Logger
-import org.elasticsearch.ElasticsearchException
-import org.elasticsearch.ExceptionsHelper
-import org.elasticsearch.action.ActionListener
-import org.elasticsearch.action.bulk.BackoffPolicy
-import org.elasticsearch.action.support.DefaultShardOperationFailedException
-import org.elasticsearch.client.ElasticsearchClient
 import org.elasticsearch.cluster.metadata.IndexMetadata
-import org.elasticsearch.common.xcontent.ToXContent
-import org.elasticsearch.common.xcontent.XContentHelper
-import org.elasticsearch.common.xcontent.XContentType
-import org.elasticsearch.rest.RestStatus
-import org.elasticsearch.transport.RemoteTransportException
-import kotlin.coroutines.resume
-import kotlin.coroutines.resumeWithException
-import kotlin.coroutines.suspendCoroutine
-
-/** Convert an object to maps and lists representation */
-fun ToXContent.convertToMap(): Map<String, Any> {
-    val bytesReference = XContentHelper.toXContent(this, XContentType.JSON, false)
-    return XContentHelper.convertToMap(bytesReference, false, XContentType.JSON).v2()
-}
-
-/**
- * Retries the given [block] of code as specified by the receiver [BackoffPolicy],
- * if [block] throws an [ElasticsearchException] that is retriable (502, 503, 504).
- *
- * If all retries fail the final exception will be rethrown. Exceptions caught during intermediate retries are
- * logged as warnings to [logger]. Similar to [org.elasticsearch.action.bulk.Retry], except this retries on
- * 502, 503, 504 error codes as well as 429.
- *
- * @param logger - logger used to log intermediate failures
- * @param retryOn - any additional [RestStatus] values that should be retried
- * @param block - the block of code to retry. This should be a suspend function.
- */
-suspend fun <T> BackoffPolicy.retry(
-    logger: Logger,
-    retryOn: List<RestStatus> = emptyList(),
-    block: suspend () -> T
-): T {
-    val iter = iterator()
-    do {
-        try {
-            return block()
-        } catch (e: ElasticsearchException) {
-            if (iter.hasNext() && (e.isRetryable() || retryOn.contains(e.status()))) {
-                val backoff = iter.next()
-                logger.warn("Operation failed. Retrying in $backoff.", e)
-                delay(backoff.millis)
-            } else {
-                throw e
-            }
-        }
-    } while (true)
-}
-
-/**
- * Retries on 502, 503 and 504 per elastic client's behavior: https://github.com/elastic/elasticsearch-net/issues/2061
- * 429 must be retried manually as it's not clear if it's ok to retry for requests other than Bulk requests.
- */
-fun ElasticsearchException.isRetryable(): Boolean {
-    return (status() in listOf(RestStatus.BAD_GATEWAY, RestStatus.SERVICE_UNAVAILABLE, RestStatus.GATEWAY_TIMEOUT))
-}
-
-/**
- * Converts [ElasticsearchClient] methods that take a callback into a kotlin suspending function.
- *
- * @param block - a block of code that is passed an [ActionListener] that should be passed to the ES client API.
- */
-suspend fun <C : ElasticsearchClient, T> C.suspendUntil(block: C.(ActionListener<T>) -> Unit): T =
-        suspendCoroutine { cont ->
-            block(object : ActionListener<T> {
-                override fun onResponse(response: T) = cont.resume(response)
-
-                override fun onFailure(e: Exception) = cont.resumeWithException(e)
-            })
-        }
-
-/**
- * Converts [LockService] methods that take a callback into a kotlin suspending function.
- *
- * @param block - a block of code that is passed an [ActionListener] that should be passed to the LockService API.
- */
-suspend fun <T> LockService.suspendUntil(block: LockService.(ActionListener<T>) -> Unit): T =
-    suspendCoroutine { cont ->
-        block(object : ActionListener<T> {
-            override fun onResponse(response: T) = cont.resume(response)
-
-            override fun onFailure(e: Exception) = cont.resumeWithException(e)
-        })
-    }
 
 /**
  * Compares current and previous IndexMetaData to determine if we should create [ManagedIndexConfig].
@@ -189,14 +97,4 @@ fun IndexMetadata.getManagedIndexMetaData(): ManagedIndexMetaData? {
         return ManagedIndexMetaData.fromMap(existingMetaDataMap)
     }
     return null
-}
-
-fun Throwable.findRemoteTransportException(): RemoteTransportException? {
-    if (this is RemoteTransportException) return this
-    return this.cause?.findRemoteTransportException()
-}
-
-fun DefaultShardOperationFailedException.getUsefulCauseString(): String {
-    val rte = this.cause?.findRemoteTransportException()
-    return if (rte == null) this.toString() else ExceptionsHelper.unwrapCause(rte).toString()
 }
